Extract estimate file DB helpers
